기타: 사용자 간 관계와 관련된 함수들을 라이브러리로 분리
